+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #106 Test API: Table.delete().where(..)
 * #107 Test API - better SELECT builder
 * #108 MySQL 8 requires special treatment of local timestamp and time
+* #109 Junit5 : lazy init of the test DataSource
 
 ## 2.0.M1